+++ src/services/product.test.ts
@@ -15,6 +15,8 @@ import {
 	StateDraft,
 	TaxCategory,
 	TaxCategoryDraft,
+	Type,
+	TypeDraft,
 } from "@commercetools/platform-sdk";
 import assert from "assert";
 import supertest from "supertest";
@@ -79,6 +81,30 @@ const productState2Draft: StateDraft = {
 	description: {
 		"nl-NL": "Product another state",
 	},
+};
+
+const productPriceTypeDraft: TypeDraft = {
+	key: "product-price",
+	name: {
+		en: "ProductPriceType",
+	},
+	description: {
+		en: "Product Price Type",
+	},
+	resourceTypeIds: ["product-price"],
+	fieldDefinitions: [
+		{
+			name: "lastModifiedAt",
+			label: {
+				en: "Last modified at",
+			},
+			required: false,
+			type: {
+				name: "DateTime",
+			},
+			inputHint: "SingleLine",
+		},
+	],
 };
 
 const publishedProductDraft: ProductDraft = {
@@ -253,6 +279,7 @@ let taxCategory1: TaxCategory;
 let taxCategory2: TaxCategory;
 let productState1: State;
 let productState2: State;
+let productPriceType: Type;
 
 async function beforeAllProductTests(mock: CommercetoolsMock) {
 	let response;
@@ -311,6 +338,12 @@ async function beforeAllProductTests(mock: CommercetoolsMock) {
 
 	expect(response.status).toBe(201);
 	productState2 = response.body;
+
+	response = await supertest(mock.app)
+		.post("/dummy/types")
+		.send(productPriceTypeDraft);
+	expect(response.status).toBe(201);
+	productPriceType = response.body;
 }
 
 describe("Product", () => {
@@ -1422,6 +1455,40 @@ describe("Product update actions", () => {
 		});
 	});
 
+	test("setProductPriceCustomField", async () => {
+		assert(productPublished, "product not created");
+		const priceId =
+			productPublished?.masterData.current.masterVariant.prices?.[0].id;
+		assert(priceId);
+
+		const response = await supertest(ctMock.app)
+			.post(`/dummy/products/${productPublished.id}`)
+			.send({
+				version: 1,
+				actions: [
+					{
+						action: "setProductPriceCustomType",
+						priceId,
+						type: {
+							typeId: "type",
+							key: productPriceType.key,
+						},
+					},
+					{
+						action: "setProductPriceCustomField",
+						name: "myCustomField",
+						value: "MyRandomValue",
+						priceId,
+					},
+				],
+			});
+		expect(response.status).toBe(200);
+		expect(
+			response.body.masterData.staged.masterVariant.prices?.[0].custom.fields
+				?.myCustomField,
+		).toBe("MyRandomValue");
+	});
+
 	// Test the general product search implementation
 	describe("Product Search - Generic", () => {
 		test("Pagination", async () => {
@@ -1466,7 +1533,7 @@ describe("Product update actions", () => {
 				expect(pagedSearchResponse.offset).toBe(24);
 				expect(pagedSearchResponse.total).toBeGreaterThan(0);
 
-				// No results, since we start at offset 24
+				// No results, since we start at offset 2400
 				const results: ProductSearchResult[] = pagedSearchResponse.results;
 				expect(results).toBeDefined();
 				expect(results.length).toBe(0);