+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ## 2.23.1
 
+## 2.29.0
+
+### Minor Changes
+
+- [#199](https://github.com/labd/commercetools-node-mock/pull/199) [`be62944`](https://github.com/labd/commercetools-node-mock/commit/be62944a5e47e03ce9ed59ea24cfee790d95320e) Thanks [@leongraumans](https://github.com/leongraumans)! - add support for "setProductPriceCustomField" product update action
+
+## 2.28.1
+
+### Patch Changes
+
+- [#204](https://github.com/labd/commercetools-node-mock/pull/204) [`a4e5a6c`](https://github.com/labd/commercetools-node-mock/commit/a4e5a6c4502e353b607e1fd3c854ef78c52951fc) Thanks [@yugaWicaksono](https://github.com/yugaWicaksono)! - Added missing test to cover OrderSetDeliveryCustomFieldAction in the form of method setDeliveryCustomField inside of OrderUpdateHandler
+
 ## 2.28.0
 
 ### Minor Changes